My first official Flutter app. About watermelon. So this is where it begins. Another journey, another project, another story to tell.
I decided, after two days of working on this project, it was finally a good time to open a repository so I can keep track of the progress I have on this project. With a good commit history and proper branching, I think this can also tell a story as good as my C# Chess project did.

Either way, I'm excited to see where this project ends up.

Can't wait!