Got the diary entry `Loader` working. Before I sleep, I'll quickly write this message.
Yes, the UI is incredibly ugly. I just needed a way to put something on the screen to validate that I have the correct input, and everything turned out okay. The buttons are big and ugly, but with some formatting, I'll make them match the schema. I'll also look for a way to hook those up to some kind of notifier, so that I can change the journal entry being edited when you select an entry in the `onPressed` argument of each `JournalCard`.

Maybe I'll also add workspaces like in Visual Studio Code. I think that would be a cool feature.

Big progress as of the past two days. Excited to see where this goes.

Also, the SQL file included in this repository is for experimental purposes only.